moved authorisation views into a stand alone module
