+++ lux/extensions/rest/views.py
@@ -1,32 +1,18 @@
-import json
-
 import lux
-from lux import route
-from lux.forms import Form
 
-from pulsar import Http404, MethodNotAllowed, BadRequest
+from pulsar import BadRequest
 from pulsar.apps.wsgi import Json
 
-from .forms import CreateUserForm, ChangePasswordForm
-from .user import AuthenticationError, logout, READ
+from .user import READ
 from .models import RestModel
-from .html import (Login, LoginPost, Logout, SignUp, ProcessLoginMixin,
-                   ForgotPassword, ComingSoon)
 from .permissions import ColumnPermissionsMixin
 
 
-__all__ = ['RestRoot', 'RestRouter', 'RestMixin', 'Authorization',
-           #
-           'Login', 'LoginPost', 'Logout', 'SignUp', 'ProcessLoginMixin',
-           'ForgotPassword', 'ComingSoon']
+__all__ = ['RestRoot', 'RestRouter', 'RestMixin']
+
 
 REST_CONTENT_TYPES = ['application/json']
 DIRECTIONS = ('asc', 'desc')
-
-
-def action(f):
-    f.is_action = True
-    return f
 
 
 class RestRoot(lux.Router):
@@ -204,105 +190,3 @@ class RestRouter(RestMixin, lux.Router):
         '''
         request.app.fire('on_preflight', request)
         return request.response
-
-
-class Authorization(RestRouter, ProcessLoginMixin):
-    '''Authentication views for
-
-    * login
-    * logout
-    * signup
-    * password change
-    * password recovery
-
-    All views respond to POST requests
-    '''
-    _model = RestModel('authorization')
-    create_user_form = CreateUserForm
-    change_password_form = ChangePasswordForm
-
-    def head(self, request):
-        return request.response
-
-    @route('/<action>', method=('post', 'options'))
-    def auth_action(self, request):
-        '''Post actions
-        '''
-        action = request.urlargs['action']
-        method = getattr(self, action, None)
-        if not getattr(method, 'is_action', False):
-            raise Http404
-
-        if request.method == 'OPTIONS':
-            request.app.fire('on_preflight', request)
-            return request.response
-
-        return method(request)
-
-    @action
-    def logout(self, request):
-        # make sure csrf is called
-        return logout(request)
-
-    @action
-    def signup(self, request):
-        '''Handle signup post data
-
-        If :attr:`.create_user_form` form is None, raise a 404 error.
-
-        A succesful response is returned by the backend
-        :meth:`.signup_response` method.
-        '''
-        if not self.create_user_form:
-            raise Http404
-
-        user = request.cache.user
-        if user.is_authenticated():
-            raise MethodNotAllowed
-
-        form = self.create_user_form(request, data=request.body_data())
-
-        if form.is_valid():
-            data = form.cleaned_data
-            auth_backend = request.cache.auth_backend
-            try:
-                user = auth_backend.create_user(request, **data)
-                return auth_backend.signup_response(request, user)
-            except AuthenticationError as e:
-                form.add_error_message(str(e))
-        return Json(form.tojson()).http_response(request)
-
-    @action
-    def change_password(self, request):
-        '''Change user password
-        '''
-        # Set change_password_form to None to remove support
-        # for password change
-        if not self.change_password_form:
-            raise Http404
-
-        user = request.cache.user
-        if not user.is_authenticated():
-            raise MethodNotAllowed
-
-        form = self.change_password_form(request, data=request.body_data())
-
-        if form.is_valid():
-            auth_backend = request.cache.auth_backend
-            password = form.cleaned_data['password']
-            auth_backend.set_password(request, user, password)
-            return auth_backend.password_changed_response(request, user)
-        return Json(form.tojson()).http_response(request)
-
-    @action
-    def dismiss_message(self, request):
-        app = request.app
-        if not app.config['SESSION_MESSAGES']:
-            raise Http404
-        session = request.cache.session
-        form = Form(request, data=request.body_data())
-        data = form.rawdata['message']
-        body = {'success': session.remove_message(data)}
-        response = request.response
-        response.content = json.dumps(body)
-        return response
